requirements.txt
pillow
psutil
numpy
